Refactor code with the latest changes applied to `trunk`
