+++ lib/experiments-page.php
@@ -114,14 +114,14 @@ function gutenberg_initialize_experiments_settings() {
 	);
 
 	add_settings_field(
-		'gutenberg-interactivity-api-navigation-block',
-		__( 'Navigation block', 'gutenberg' ),
+		'gutenberg-interactivity-api-core-blocks',
+		__( 'Core blocks', 'gutenberg' ),
 		'gutenberg_display_experiment_field',
 		'gutenberg-experiments',
 		'gutenberg_experiments_section',
 		array(
-			'label' => __( 'Test the Navigation block using the Interactivity API', 'gutenberg' ),
-			'id'    => 'gutenberg-interactivity-api-navigation-block',
+			'label' => __( 'Test the core blocks using the Interactivity API', 'gutenberg' ),
+			'id'    => 'gutenberg-interactivity-api-core-blocks',
 		)
 	);
