+++ packages/block-library/src/file/interactivity.js
@@ -1,8 +1,8 @@
 /**
  * Internal dependencies
  */
-import { store } from '../../utils/interactivity';
-import { browserSupportsPdfs } from '../utils';
+import { store } from '../utils/interactivity';
+import { browserSupportsPdfs } from './utils';
 
 store( {
 	effects: {
